% Auto-generated: do not edit by hand
\name{''SigmaJSComponent}

\alias{''SigmaJSComponent}

\title{SigmaJSComponent component}

\description{
Component to use SigmaJS renderer in a React/Dash application
}

\usage{
''SigmaJSComponent(id=NULL, className=NULL, graph_data=NULL, label=NULL,
settings=NULL, style=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. The class(es) used to identify this component.}

\item{graph_data}{Named list. Dict/JSON that contains graph definition. See https://graphology.github.io/serialization.html#format}

\item{label}{Character. Label}

\item{settings}{Named list. Settings}

\item{style}{Named list. Inline Style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

